+++ app/controllers/admin/customer_credits_controller.rb
@@ -13,7 +13,7 @@ module Admin
 
       if @credit.save
         flash['success'] = t('CreateSuccess')
-        CustomerMailer.with(customer: @credit.customer, gift_amount: @credit.balance).gift_card.deliver_now
+        CustomerMailer.with(customer: @credit.customer, gift_amount: @credit.balance).gift_card.deliver_later
         redirect_to admin_customer_path @credit.customer
       else
         render 'new'