+++ app/workers/payment_deadline_reminder_worker.rb
@@ -1,0 +1,7 @@
+# frozen_string_literal: true
+
+class PaymentDeadlineReminderWorker
+  include Sidekiq::Worker
+
+  def perform; end
+end