+++ spec/factories/refunds.rb
@@ -4,6 +4,6 @@ FactoryBot.define do
   factory :refund do
     invoice
     customer_credit
-    amount { Faker::Number.decimal(2) }
+    amount { Faker::Commerce.material }
   end
 end