+++ spec/factories/workshop_types.rb
@@ -9,7 +9,7 @@ FactoryBot.define do
     default_reservation_allow { true }
     default_reservation_allow_multiple { true }
     default_reservation_cancel_minimum_not_met { true }
-    default_reservation_price { Faker::Number.decimal(2) }
+    default_reservation_price { Faker::Commerce.material }
     default_reservation_maximum { 18 }
     default_reservation_minimum { Faker::Number.between(1, default_reservation_maximum) }